Move Container to separate folder
