+++ src/pages/Dashboard/Container/NavigationButton.tsx
@@ -1,5 +1,5 @@
 import { Dialog } from "@headlessui/react";
-import { LinkGroup } from "./types";
+import { LinkGroup } from "../types";
 import { useModalContext } from "contexts/ModalContext";
 import Sidebar from "./Sidebar";
